Move types to common
