+++ front/src/common/types.ts
