+++ front/src/multiplication/game_logic.ts
@@ -1,4 +1,4 @@
-import * as Model from './types';
+import * as Model from '../common/types';
 import * as DifficultyLevels from './difficulty_levels';
 import * as QuestionGeneration from './questions_generator';
